Break reference integration DTO import cycle
