+++ src/services/query_service/app/dtos/reference_integration_benchmark_market_series_dto.py
@@ -6,8 +6,8 @@ from typing import Literal
 
 from pydantic import BaseModel, ConfigDict, Field, model_validator
 
+from .reference_integration_common_dto import IntegrationWindow
 from .reference_integration_dto import (
-    IntegrationWindow,
     ReferencePageMetadata,
     ReferencePageRequest,
     SeriesRequest,
