+++ src/services/query_service/app/dtos/reference_integration_performance_component_economics_dto.py
@@ -6,7 +6,7 @@ from typing import Literal
 
 from pydantic import BaseModel, ConfigDict, Field, model_validator
 
-from .reference_integration_dto import IntegrationWindow
+from .reference_integration_common_dto import IntegrationWindow
 from .source_data_product_identity import (
     SourceDataProductRuntimeMetadata,
     product_name_field,
